🔥 remove #259: RoundedLabel 파일 분리
